chore: merge tool_runner into utils
